+++ src/settings/FileExtensions/FileExtensions.js
@@ -34,7 +34,6 @@ import {
 } from '../../components';
 import { ViewFileExtension } from './ViewFileExtension';
 import { SettingPage } from '../SettingPage';
-import { ENTITY_CONFIGS } from '../../utils/constants';
 
 const INITIAL_RESULT_COUNT = 30;
 const RESULT_COUNT_INCREMENT = 30;
@@ -270,7 +269,6 @@ export class FileExtensions extends Component {
 
     const urlQuery = queryString.parse(location.search);
     const searchTerm = trimSearchTerm(urlQuery.query);
-    const { ENTITY_KEY } = ENTITY_CONFIGS.FILE_EXTENSIONS;
 
     return (
       <IntlConsumer>
@@ -298,7 +296,10 @@ export class FileExtensions extends Component {
                   resultsLabel={label}
                   defaultSort="extension"
                   actionMenu={this.renderActionMenu}
-                  resultsFormatter={listTemplate(ENTITY_KEY, searchTerm)}
+                  resultsFormatter={listTemplate({
+                    intl,
+                    searchTerm,
+                  })}
                   visibleColumns={this.visibleColumns}
                   columnMapping={{
                     extension: intl.formatMessage({ id: 'ui-data-import.settings.fileExtension.extension' }),