+++ src/settings/MatchProfiles/MatchProfiles.js
@@ -267,7 +267,12 @@ export class MatchProfiles extends Component {
                 resultsLabel={label}
                 defaultSort="name"
                 actionMenu={this.renderActionMenu}
-                resultsFormatter={listTemplate(ENTITY_KEY, searchTerm, selectRecord, selectedRecords)}
+                resultsFormatter={listTemplate({
+                  entityKey: ENTITY_KEY,
+                  searchTerm,
+                  selectRecord,
+                  selectedRecords,
+                })}
                 visibleColumns={this.visibleColumns}
                 columnMapping={{
                   selected: (