+++ test/bigtest/interactors/match-profiles/match-profile-details.js
@@ -13,6 +13,7 @@ import { CheckboxInteractor } from '../checkbox-interactor';
   list = new MultiColumnListInteractor('#associated-job-profiles-list');
   selectAllCheckBox = new CheckboxInteractor('[data-test-select-all-associated-job-profiles-checkbox]');
   checkBoxes = collection('[data-test-select-item]', CheckboxInteractor);
+  jobProfilesLinks = collection('[data-test-job-profile-link]', ButtonInteractor);
 }
 
 @interactor class MatchProfileDetailsInteractor {